hack the broken WebRequest.AddRange, add down test

+++ QiniuFSTest/010_IOTest.fs
@@ -13,24 +13,6 @@ open TestBase
 
 [<TestFixture>]
 type IOTest() =
-    let smallName = "small.dat"
-    let smallPath = Path.Combine(testPath, smallName)
-    do genNotExist smallPath (1 <<< 20) // 1M
-    let smallMD5 = using (File.OpenRead smallPath) md5
-
-    let bigName = "big.dat"
-    let bigPath = Path.Combine(testPath, bigName)
-    do genNotExist bigPath (1 <<< 23) // 8M
-    let bigMD5 = using (File.OpenRead(bigPath)) md5
-
-    member this.Token(key : String) =
-        let policy = { 
-            IO.putPolicy with
-                scope = IO.scope <| entry tc.BUCKET key
-                deadline = IO.defaultDeadline()
-        }
-        IO.sign c policy
-
     member this.GetMD5AndDelete(key : String) =
         let url = IO.publicUrl tc.DOMAIN key
         let req = WebRequest.Create url :?> HttpWebRequest
@@ -43,7 +25,7 @@ type IOTest() =
     [<Test>]
     member this.PutTest() =
         let key = String.Format("{0}_{1}", ticks(), smallName)
-        let token = this.Token key
+        let token = uptoken key
         use stream = File.OpenRead smallPath
         IO.put c token key stream IO.putExtra |> checkSynchro
         Assert.AreEqual(smallMD5, this.GetMD5AndDelete key)
@@ -51,7 +33,7 @@ type IOTest() =
     [<Test>]
     member this.PutCrcTest() =
         let key = String.Format("{0}_{1}", ticks(), smallName)
-        let token = this.Token key
+        let token = uptoken key
         use stream = File.OpenRead smallPath
         let extra = { IO.putExtra with checkCrc = IO.CheckCrc.Auto }
         IO.put c token key stream extra |> checkSynchro
@@ -60,27 +42,28 @@ type IOTest() =
     [<Test>]
     member this.RPutTest() =
         let key = String.Format("{0}_{1}", ticks(), bigName)
-        let token = this.Token key
+        let token = uptoken key
         use stream = File.OpenRead bigPath
         RIO.rput c token key stream RIO.rputExtra |> checkSynchro
         Assert.AreEqual(bigMD5, this.GetMD5AndDelete key)
 
     [<Test>]
     member this.ResumebleRPutTest() =
-        let progressesData = "progresses.dat"
+        let progressesData = "rputProgresses.dat"
         let progressesPath = Path.Combine(testPath, progressesData)
-        if File.Exists(progressesPath) then
-            File.Delete(progressesPath)
+        if File.Exists(progressesPath) then File.Delete(progressesPath)
 
         let key = String.Format("{0}_{1}", ticks(), bigName)
-        let token = this.Token key
+        let token = uptoken key
         let notifyCancelCount = 10
 
         let upload _ =    
             let progresses =
                 if File.Exists progressesPath then
                     use data = File.OpenRead(progressesPath)
-                    readJsons<RIO.Progress> data |> Seq.toArray
+                    readJsons<RIO.Progress> data 
+                    |> Seq.toArray 
+                    |> RIO.cleanProgresses
                 else Array.zeroCreate<RIO.Progress> 0
             
             use data = File.Create(progressesPath)
@@ -89,13 +72,11 @@ type IOTest() =
             let cs = new CancellationTokenSource()
             
             let notifyCount = ref -1
-            let notifyLock = new Object()
-            let notify (p : RIO.Progress) =
-                lock notifyLock (fun _ -> 
-                    incr notifyCount
-                    writeJson data p
-                    if unref notifyCount = notifyCancelCount then
-                        cs.Cancel())
+            let notify (p : RIO.Progress) =  
+                incr notifyCount
+                writeJson data p
+                if unref notifyCount = notifyCancelCount then
+                    cs.Cancel()
             
             let extra = {
                 RIO.rputExtra with
@@ -103,7 +84,9 @@ type IOTest() =
                     RIO.notify = notify
             }
             use stream = File.OpenRead bigPath
+
             let work = RIO.rput c token key stream extra 
+
             try Async.RunSynchronously(work, -1, cs.Token)
             with | :? OperationCanceledException -> 
                 IO.PutError({ error = "Upload not done yet and just try again"})
